requirements.txt
-i https://pypi.org/simple
aiohappyeyeballs==2.4.0; python_version >= '3.8'
aiohttp==3.10.5; python_version >= '3.8'
aiohttp-retry==2.8.3; python_version >= '3.7'
aiosignal==1.3.1; python_version >= '3.7'
aiosmtplib==3.0.2; python_version >= '3.8'
alembic==1.13.2; python_version >= '3.8'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.4.0; python_version >= '3.8'
asyncpg==0.29.0; python_full_version >= '3.8.0'
attrs==24.2.0; python_version >= '3.7'
bcrypt==3.2.0; python_version >= '3.6'
certifi==2024.8.30; python_version >= '3.6'
cffi==1.17.1; python_version >= '3.8'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
colorama==0.4.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
cryptography==43.0.3; python_version >= '3.7'
dnspython==2.6.1; python_version >= '3.8'
email-validator==2.2.0; python_version >= '3.8'
fastapi[standard]==0.115.2; python_version >= '3.8'
fastapi-cli[standard]==0.0.5; python_version >= '3.8'
frozenlist==1.4.1; python_version >= '3.8'
greenlet==3.1.0; python_version >= '3.7'
h11==0.14.0; python_version >= '3.7'
http-ece==1.2.1
httpcore==1.0.5; python_version >= '3.8'
httptools==0.6.1; python_full_version >= '3.8.0'
httpx==0.27.2; python_version >= '3.8'
idna==3.8; python_version >= '3.6'
jinja2==3.1.4; python_version >= '3.7'
joblib==1.4.2; python_version >= '3.8'
mako==1.3.5; python_version >= '3.8'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==2.1.5; python_version >= '3.7'
mdurl==0.1.2; python_version >= '3.7'
multidict==6.1.0; python_version >= '3.8'
numpy==2.1.2; python_version >= '3.12'
pandas==2.2.3; python_version >= '3.9'
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.9; python_version >= '3.7'
py-vapid==1.9.1
pycparser==2.22; python_version >= '3.8'
pydantic[dotenv,email]==2.9.1; python_version >= '3.8'
pydantic-core==2.23.3; python_version >= '3.8'
pydantic-settings==2.5.2; python_version >= '3.8'
pygments==2.18.0; python_version >= '3.8'
pyjwt==2.9.0; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==1.0.1; python_version >= '3.8'
python-http-client==3.3.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-multipart==0.0.9; python_version >= '3.8'
pytz==2024.2
pywebpush==2.0.1
pyyaml==6.0.2; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
rich==13.8.1; python_full_version >= '3.7.0'
scikit-learn==1.5.2; python_version >= '3.9'
scipy==1.14.1; python_version >= '3.10'
sendgrid==6.11.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
shellingham==1.5.4; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy[asyncio]==2.0.34; python_version >= '3.7'
starkbank-ecdsa==2.2.0
starlette==0.38.5; python_version >= '3.8'
threadpoolctl==3.5.0; python_version >= '3.8'
twilio==9.3.0; python_full_version >= '3.7.0'
typer==0.12.5; python_version >= '3.7'
typing-extensions==4.12.2; python_version >= '3.8'
tzdata==2024.2; python_version >= '2'
urllib3==2.2.3; python_version >= '3.8'
uvicorn[standard]==0.30.6; python_version >= '3.8'
watchfiles==0.24.0; python_version >= '3.8'
websockets==13.0.1; python_version >= '3.8'
yarl==1.11.1; python_version >= '3.8'
